server/requirements.txt
channels==2.4.0
channels-redis==3.0.1
Django==3.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
Pillow==7.2.0
pytest==6.0.1
pytest-asyncio==0.14.0
pytest-django==3.9.0
psycopg2-binary==2.8.5
